+++ ascii-th/library/ASCII/QuasiQuoters.hs
@@ -9,10 +9,13 @@ module ASCII.QuasiQuoters
     char,
     string,
     caseless,
+    upper,
+    lower,
+    ofCase,
   )
   where
 
-import ASCII.Case (Case)
+import ASCII.Case (Case (..))
 import ASCII.Caseless (CaselessChar)
 import ASCII.Char (Char)
 import Control.Monad (return, (>=>))
@@ -22,9 +25,9 @@ import Data.Maybe (Maybe (..))
 import Language.Haskell.TH.Quote (QuasiQuoter (..))
 import Language.Haskell.TH.Syntax (Exp, Pat, Q)
 
-import qualified ASCII.TemplateHaskell as TH
-import qualified ASCII.Caseless as Caseless
+import qualified ASCII.Case as Case
 import qualified ASCII.Superset as S
+import qualified ASCII.TemplateHaskell as TH
 import qualified Data.Char as Unicode
 import qualified Data.List as List
 import qualified Data.String as Unicode
@@ -123,23 +126,70 @@ A case-insensitive match of any type belonging to the
 caseless :: QuasiQuoter
 caseless = expPatQQ requireAsciiListCI TH.caselessListExp TH.caselessIsStringPat
 
+-- | @'ofCase' 'LowerCase'@
+lower :: QuasiQuoter
+lower = ofCase LowerCase
+
+-- | @'ofCase' 'UpperCase'@
+upper :: QuasiQuoter
+upper = ofCase UpperCase
+
+{-| An expression or pattern corresponding to an ASCII string where all the
+letters are of a particular case
+
+The letters in the body of the quasi-quotation may be written in any case
+you like; they will be converted to the designated case automatically.
+
+Since functions cannot be used as quasi-quoters, you likely want 'lower' or
+'upper' instead of using this function directly.
+
+=== In an expression context
+
+The expression can become any type belonging to the 'ASCII.Superset.FromString'
+class. Any letters in the quoted content will be converted to the given 'Case'.
+
+=== In a pattern context
+
+The pattern matches a value of a type satisfying the 'ASCII.Superset.ToString'
+constraint. A value matches this pattern if:
+
+* All of the letters in the tested value are in the given 'Case'
+* The tested value satisfies a case-insensitive comparison with the
+  quasi-quoted content
+
+-}
+ofCase :: Case -> QuasiQuoter
+ofCase c = expPatQQ (requireAsciiToCase c) TH.isStringExp TH.isStringPat
+
+{-| Require the string to consist of exactly one ASCII character -}
 requireOneAscii :: Unicode.String -> Q Char
 requireOneAscii = requireOne >=> requireAscii
 
+{-| Require the list to consist of exactly one element -}
 oneMaybe :: [a] -> Maybe a
 oneMaybe xs = case xs of [x] -> Just x; _ -> Nothing
 
+{-| Require the string to consist of exactly one character -}
 requireOne :: Unicode.String -> Q Unicode.Char
 requireOne = oneMaybe || "Must be exactly one character."
 
+{-| Require the character to be ASCII -}
 requireAscii :: Unicode.Char -> Q Char
 requireAscii = S.toCharMaybe || "Must be an ASCII character."
 
+{-| Require the string to consist of all ASCII characters -}
 requireAsciiList :: Unicode.String -> Q [Char]
 requireAsciiList = S.toCharListMaybe || "Must be only ASCII characters."
 
+{-| Require the string to consist of all ASCII characters,
+and return them with letter case discarded -}
 requireAsciiListCI :: Unicode.String -> Q [CaselessChar]
 requireAsciiListCI = S.toCaselessCharListMaybe || "Must be only ASCII characters."
+
+{-| Require the string to consist of all ASCII characters,
+and return them converted to the given case -}
+requireAsciiToCase :: Case -> Unicode.String -> Q [Char]
+requireAsciiToCase c s = List.map (Case.toCase c) <$> requireAsciiList s
 
 (||) :: (a -> Maybe b) -> Unicode.String -> a -> Q b
 f || msg = \a -> case f a of Just b -> return b; Nothing -> fail msg