+++ ascii-th/changelog.md
@@ -3,6 +3,9 @@
 New in `ASCII.QuasiQuoters`:
 
 - `caseless`
+- `upper`
+- `lower`
+- `ofCase`
 
 New in `ASCII.TemplateHaskell`: